+++ tests/codesync.test.js
@@ -3,11 +3,18 @@
  */
 
 import fs from "fs";
+import path from "path";
 import yaml from "js-yaml";
 import untildify from "untildify";
 
 import extension from "../lib/codesync";
-import {COMMAND, getRepoInSyncMsg, NOTIFICATION} from "../lib/constants";
+import {
+    getRepoInSyncMsg,
+    getDirectoryIsSyncedMsg,
+    getDirectorySyncIgnoredMsg, 
+    NOTIFICATION, 
+    SYNCIGNORE
+} from "../lib/constants";
 import {
     SignUpHandler,
     SyncHandler,
@@ -21,19 +28,15 @@ import {
     buildAtomEnv,
     Config,
     getConfigFilePath,
-    getUserFilePath,
     randomBaseRepoPath,
     randomRepoPath
 } from "./helpers/helpers";
 import {logout} from "../lib/utils/auth_utils";
-import {UnsyncRepoView} from "../lib/views";
-
+import { CodeSyncState, CODESYNC_STATES } from "../lib/utils/state_utils";
 
 describe("Extension",() => {
     const baseRepoPath = randomBaseRepoPath();
     const repoPath = randomRepoPath();
-    const userFilePath = getUserFilePath(baseRepoPath);
-    const userData = {"dummy_email": {access_token: "ABC"}};
     const configPath = getConfigFilePath(baseRepoPath);
     const configData = {repos: {}};
     configData.repos[repoPath] = {branches: {}};
@@ -43,8 +46,8 @@ describe("Extension",() => {
         buildAtomEnv();
         untildify.mockReturnValue(baseRepoPath);
         createSystemDirectories();
+        fs.mkdirSync(repoPath, {recursive: true});
         global.IS_CODESYNC_DEV = true;
-        fs.mkdirSync(baseRepoPath, { recursive: true });
         extension.consumeStatusBar({
             addRightTile: jest.fn(),
             getRightTiles: jest.fn(() => []),
@@ -54,6 +57,7 @@ describe("Extension",() => {
     });
 
     afterEach(() => {
+        fs.rmSync(repoPath, { recursive: true, force: true });
         fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
@@ -63,7 +67,7 @@ describe("Extension",() => {
         expect(atom.menu.add).toHaveBeenCalledTimes(1);
         expect(atom.contextMenu.add).toHaveBeenCalledTimes(1);
         expect(atom.views.addViewProvider).toHaveBeenCalledTimes(1);
-        expect(atom.commands.add).toHaveBeenCalledTimes(6);
+        expect(atom.commands.add).toHaveBeenCalledTimes(7);
 
         // Register commands
         expect(atom.commands.add.mock.calls[0][0]).toStrictEqual('atom-workspace');
@@ -112,7 +116,7 @@ describe("Extension",() => {
         expect(atom.menu.add).toHaveBeenCalledTimes(1);
         expect(atom.contextMenu.add).toHaveBeenCalledTimes(1);
         expect(atom.views.addViewProvider).toHaveBeenCalledTimes(1);
-        expect(atom.commands.add).toHaveBeenCalledTimes(6);
+        expect(atom.commands.add).toHaveBeenCalledTimes(7);
         // Should show Welcome msg
         expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
         expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(NOTIFICATION.WELCOME_MSG);
@@ -123,13 +127,13 @@ describe("Extension",() => {
     });
 
     test("With user, repo not synced", async () => {
-        fs.writeFileSync(userFilePath, yaml.safeDump(userData));
+        addUser(baseRepoPath);
         atom.project.getPaths.mockReturnValue([repoPath]);
         await extension.activate({});
         expect(atom.menu.add).toHaveBeenCalledTimes(1);
         expect(atom.contextMenu.add).toHaveBeenCalledTimes(1);
         expect(atom.views.addViewProvider).toHaveBeenCalledTimes(1);
-        expect(atom.commands.add).toHaveBeenCalledTimes(6);
+        expect(atom.commands.add).toHaveBeenCalledTimes(7);
 
         expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
         expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(NOTIFICATION.CONNECT_REPO);
@@ -138,11 +142,12 @@ describe("Extension",() => {
         expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.CONNECT);
         expect(options.buttons[1].text).toStrictEqual(NOTIFICATION.IGNORE);
         expect(options.dismissable).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(false);
     });
 
     test("With user, repo is disconnected", async () => {
         atom.project.getPaths.mockReturnValue([repoPath]);
-        fs.writeFileSync(userFilePath, yaml.safeDump(userData));
+        addUser(baseRepoPath);
         const _configData = JSON.parse(JSON.stringify(configData));
         _configData.repos[repoPath].is_disconnected = true;
         fs.writeFileSync(configPath, yaml.safeDump(_configData));
@@ -150,7 +155,7 @@ describe("Extension",() => {
         expect(atom.menu.add).toHaveBeenCalledTimes(1);
         expect(atom.contextMenu.add).toHaveBeenCalledTimes(1);
         expect(atom.views.addViewProvider).toHaveBeenCalledTimes(1);
-        expect(atom.commands.add).toHaveBeenCalledTimes(6);
+        expect(atom.commands.add).toHaveBeenCalledTimes(7);
         expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
         expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(NOTIFICATION.CONNECT_REPO);
         const options = atom.notifications.addInfo.mock.calls[0][1];
@@ -158,6 +163,7 @@ describe("Extension",() => {
         expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.CONNECT);
         expect(options.buttons[1].text).toStrictEqual(NOTIFICATION.IGNORE);
         expect(options.dismissable).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(false);
     });
 
     test("With user, repo is in sync", async () => {
@@ -173,5 +179,50 @@ describe("Extension",() => {
         expect(options.buttons).toHaveLength(1);
         expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.TRACK_IT);
         expect(options.dismissable).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(true);
+    });
+
+    test("With user, repo is sub directory and synced", async () => {
+        const subDirName = "directory";
+        const configUtil = new Config(repoPath, configPath);
+        configUtil.addRepo();
+        addUser(baseRepoPath);
+        const subDir = path.join(repoPath, subDirName);
+        atom.project.getPaths.mockReturnValue([subDir]);
+        await extension.activate({});
+        expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
+        const json = getDirectoryIsSyncedMsg(subDir, repoPath);
+        expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(json.msg);
+        const options = atom.notifications.addInfo.mock.calls[0][1];
+        expect(options.buttons).toHaveLength(1);
+        expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.TRACK_PARENT_REPO);
+        expect(options.dismissable).toBe(true);
+        expect(options.detail).toBe(json.detail);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SUB_DIR)).toBe(true);
+    });
+
+    test("With user, repo is sub directory and syncignored", async () => {
+        const subDirName = "directory";
+        const configUtil = new Config(repoPath, configPath);
+        configUtil.addRepo();
+        addUser(baseRepoPath);
+        // Add subDir to .syncignore
+        const syncignorePath = path.join(repoPath, SYNCIGNORE);
+        fs.writeFileSync(syncignorePath, subDirName);
+        const subDir = path.join(repoPath, subDirName);
+        atom.project.getPaths.mockReturnValue([subDir]);
+        await extension.activate({});
+        expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
+        const msg = getDirectorySyncIgnoredMsg(subDir, repoPath);
+        expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(msg);
+        const options = atom.notifications.addInfo.mock.calls[0][1];
+        expect(options.buttons).toHaveLength(3);
+        expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.OPEN_SYNCIGNORE);
+        expect(options.buttons[1].text).toStrictEqual(NOTIFICATION.TRACK_PARENT_REPO);
+        expect(options.buttons[2].text).toStrictEqual(NOTIFICATION.UNSYNC_PARENT_REPO);
+        expect(options.dismissable).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(false);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SYNCIGNORED_SUB_DIR)).toBe(true);
     });
 });