+++ tests/test_events/utils/shouldIgnoreFile.test.js
@@ -1,7 +1,7 @@
 import fs from "fs";
 import path from "path";
 
-import {shouldIgnoreFile} from "../../../lib/events/utils";
+import {shouldIgnorePath} from "../../../lib/events/utils";
 import {getSyncIgnoreFilePath, randomBaseRepoPath, randomRepoPath} from "../../helpers/helpers";
 import {IGNORABLE_DIRECTORIES} from "../../../lib/constants";
 
@@ -22,29 +22,29 @@ beforeAll(() => {
 });
 
 afterAll(() => {
-    fs.rmdirSync(baseRepoPath, { recursive: true });
-    fs.rmdirSync(repoPath, { recursive: true });
+    fs.rmSync(baseRepoPath, { recursive: true });
+    fs.rmSync(repoPath, { recursive: true });
 });
 
 test("Standard ignorable directories",  () => {
     IGNORABLE_DIRECTORIES.forEach((item) => {
-        expect(shouldIgnoreFile(repoPath, item)).toBe(true);
+        expect(shouldIgnorePath(repoPath, item)).toBe(true);
     });
 });
 
-test("shouldIgnoreFile with normal file and no .syncignore",  () => {
-    expect(shouldIgnoreFile(repoPath, normalFilePath)).toBe(false);
+test("shouldIgnorePath with normal file and no .syncignore",  () => {
+    expect(shouldIgnorePath(repoPath, normalFilePath)).toBe(false);
 });
 
-test("shouldIgnoreFile with normal file and with .syncignore",  () => {
+test("shouldIgnorePath with normal file and with .syncignore",  () => {
     fs.writeFileSync(syncIgnorePath, syncIgnoreData);
-    expect(shouldIgnoreFile(repoPath, normalFilePath)).toBe(false);
+    expect(shouldIgnorePath(repoPath, normalFilePath)).toBe(false);
     fs.rmSync(syncIgnorePath);
 });
 
-test("shouldIgnoreFile with ignorable file",  () => {
+test("shouldIgnorePath with ignorable file",  () => {
     fs.writeFileSync(syncIgnorePath, syncIgnoreData);
-    expect(shouldIgnoreFile(repoPath, ignorableFilePath)).toBe(true);
+    expect(shouldIgnorePath(repoPath, ignorableFilePath)).toBe(true);
     fs.rmSync(syncIgnorePath);
 });