+++ tests/test_utils/setup_utils.test.js
@@ -1,4 +1,5 @@
 import fs from "fs";
+import path from "path";
 import yaml from "js-yaml";
 import untildify from "untildify";
 
@@ -13,9 +14,15 @@ import {
     writeFile,
     Config, addUser
 } from "../helpers/helpers";
-import {getRepoInSyncMsg, NOTIFICATION} from "../../lib/constants";
-import {createSystemDirectories, setupCodeSync, showConnectRepoView, showLogIn} from "../../lib/utils/setup_utils";
-
+import {
+    getRepoInSyncMsg,
+    getDirectorySyncIgnoredMsg,
+    getDirectoryIsSyncedMsg,
+    NOTIFICATION,
+    SYNCIGNORE} from "../../lib/constants";
+import {createSystemDirectories, setupCodeSync, showConnectRepoView, 
+    showLogIn, showRepoIsSyncIgnoredView} from "../../lib/utils/setup_utils";
+import { CodeSyncState, CODESYNC_STATES } from "../../lib/utils/state_utils";
 
 describe("createSystemDirectories",  () => {
     const baseRepoPath = randomBaseRepoPath();
@@ -57,6 +64,8 @@ describe("setupCodeSync",  () => {
         untildify.mockReturnValue(baseRepoPath);
         mkDir(baseRepoPath);
         mkDir(repoPath);
+        CodeSyncState.set(CODESYNC_STATES.REPO_IS_IN_SYNC, false);
+        CodeSyncState.set(CODESYNC_STATES.USER_EMAIL, null);
     });
 
     afterEach(() => {
@@ -99,7 +108,7 @@ describe("setupCodeSync",  () => {
 
     test('with no active user', async () => {
         addUser(baseRepoPath, false);
-        const port = await setupCodeSync(undefined);
+        const port = await setupCodeSync(repoPath);
         // should return port number
         expect(port).toBeTruthy();
         expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
@@ -112,12 +121,13 @@ describe("setupCodeSync",  () => {
 
     test('with user no repo opened', async () => {
         fs.writeFileSync(userFilePath, yaml.safeDump(userData));
-        await setupCodeSync(undefined);
+        await setupCodeSync("");
         expect(atom.notifications.addInfo).toHaveBeenCalledTimes(0);
     });
 
     test('with user and repo not synced', async () => {
         writeFile(userFilePath, yaml.safeDump(userData));
+        atom.project.getPaths.mockReturnValue([repoPath]);
         const port = await setupCodeSync(repoPath);
         // should return port number
         expect(port).toBeTruthy();
@@ -129,6 +139,8 @@ describe("setupCodeSync",  () => {
         expect(options.buttons[1].text).toStrictEqual(NOTIFICATION.IGNORE);
         expect(options.dismissable).toBe(true);
         fs.rmSync(userFilePath);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(false);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SUB_DIR)).toBe(false);
     });
 
     test('with synced repo',  async () => {
@@ -136,24 +148,106 @@ describe("setupCodeSync",  () => {
         const configUtil = new Config(repoPath, configPath);
         configUtil.addRepo();
         addUser(baseRepoPath);
+        atom.project.getPaths.mockReturnValue([repoPath]);
         const port = await setupCodeSync(repoPath);
         // should return port number
         expect(port).toBeFalsy();
         expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
-        const repoInSyncMsg = getRepoInSyncMsg(repoPath);
-        expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(repoInSyncMsg);
+        const msg = getRepoInSyncMsg(repoPath);
+        expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(msg);
         const options = atom.notifications.addInfo.mock.calls[0][1];
         expect(options.buttons).toHaveLength(1);
         expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.TRACK_IT);
         expect(options.dismissable).toBe(true);
         fs.rmSync(userFilePath);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SUB_DIR)).toBe(false);
     });
 
     test('showConnectRepoView',  async () => {
-        atom.project.getPaths.mockReturnValue([repoPath]);
         writeFile(configPath, yaml.safeDump({repos: {}}));
-        const shouldShowConnectRepoView = showConnectRepoView();
+        const shouldShowConnectRepoView = showConnectRepoView(repoPath);
         expect(shouldShowConnectRepoView).toBe(true);
+    });
+
+    test('showRepoIsSyncIgnoredView',  async () => {
+        fs.writeFileSync(configPath, yaml.safeDump({repos: {}}));
+        const shouldShow = showRepoIsSyncIgnoredView(repoPath);
+        expect(shouldShow).toBe(false);
+    });
+
+    test('with sub directory',  async () => {
+        fs.writeFileSync(userFilePath, yaml.safeDump(userData));
+        const configUtil = new Config(repoPath, configPath);
+        configUtil.addRepo();
+        addUser(baseRepoPath);
+        const subDir = path.join(repoPath, "directory");
+        const port = await setupCodeSync(subDir);
+        // should return port number
+        expect(port).toBeFalsy();
+        expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
+        const json = getDirectoryIsSyncedMsg(subDir, repoPath);
+        expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(json.msg);
+        const options = atom.notifications.addInfo.mock.calls[0][1];
+        expect(options.buttons).toHaveLength(1);
+        expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.TRACK_PARENT_REPO);
+        expect(options.dismissable).toBe(true);
+        expect(options.detail).toStrictEqual(json.detail);
+        fs.rmSync(userFilePath);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SUB_DIR)).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SYNCIGNORED_SUB_DIR)).toBe(false);
+    });
+
+    test('with sub directory syncignored',  async () => {
+        const subDirName = "directory";
+        fs.writeFileSync(userFilePath, yaml.safeDump(userData));
+        const configUtil = new Config(repoPath, configPath);
+        configUtil.addRepo();
+        addUser(baseRepoPath);
+        // Add subDir to .syncignore
+        const syncignorePath = path.join(repoPath, SYNCIGNORE);
+        fs.writeFileSync(syncignorePath, subDirName);
+        const subDir = path.join(repoPath, subDirName);
+        const port = await setupCodeSync(subDir);
+        // should return port number
+        expect(port).toBeTruthy();
+        expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
+        const msg = getDirectorySyncIgnoredMsg(subDir, repoPath);
+        expect(atom.notifications.addInfo.mock.calls[0][0]).toStrictEqual(msg);
+        const options = atom.notifications.addInfo.mock.calls[0][1];
+        expect(options.buttons).toHaveLength(3);
+        expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.OPEN_SYNCIGNORE);
+        expect(options.buttons[1].text).toStrictEqual(NOTIFICATION.TRACK_PARENT_REPO);
+        expect(options.buttons[2].text).toStrictEqual(NOTIFICATION.UNSYNC_PARENT_REPO);
+        expect(options.dismissable).toBe(true);
+        fs.rmSync(userFilePath);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(false);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SUB_DIR)).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SYNCIGNORED_SUB_DIR)).toBe(true);
+    });
+
+    test('with sub directory and parent is_disconnected',  async () => {
+        fs.writeFileSync(userFilePath, yaml.safeDump(userData));
+        const configUtil = new Config(repoPath, configPath);
+        configUtil.addRepo(true);
+        addUser(baseRepoPath);
+        const subDir = path.join(repoPath, "directory");
+        atom.project.getPaths.mockReturnValue([repoPath]);
+        const port = await setupCodeSync(subDir);
+        // should return port number
+        expect(port).toBeTruthy();
+        expect(atom.notifications.addInfo).toHaveBeenCalledTimes(1);
+        expect(atom.notifications.addInfo.mock.calls[0][0]).toBe(NOTIFICATION.CONNECT_REPO);
+        const options = atom.notifications.addInfo.mock.calls[0][1];
+        expect(options.buttons).toHaveLength(2);
+        expect(options.buttons[0].text).toStrictEqual(NOTIFICATION.CONNECT);
+        expect(options.buttons[1].text).toStrictEqual(NOTIFICATION.IGNORE);
+        expect(options.dismissable).toBe(true);
+        fs.rmSync(userFilePath);
+        expect(CodeSyncState.get(CODESYNC_STATES.REPO_IS_IN_SYNC)).toBe(false);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SUB_DIR)).toBe(true);
+        expect(CodeSyncState.get(CODESYNC_STATES.IS_SYNCIGNORED_SUB_DIR)).toBe(false);
     });
 });